@@ -1,5 +1,5 @@
 module github.com/go-cmd/cmd
 
-go 1.16
+go 1.17
 
 require github.com/go-test/deep v1.0.7
